+++ Mercury/Backtests/BacktestStrategies/CGTrend1.cs
@@ -20,12 +20,12 @@ namespace Mercury.Backtests.BacktestStrategies
 	{
 		private SmartRandom r = new();
 
-		public double RsiL = 40;
-		public double RsiH = 60;
-		public double RsiL2 = 60;
-		public double RsiH2 = 40;
+		public decimal RsiL = 40;
+		public decimal RsiH = 60;
+		public decimal RsiL2 = 60;
+		public decimal RsiH2 = 40;
 
-		public double RRange = 5;
+		public decimal RRange = 5;
 
 		/// <summary>
 		/// Interval이 다른 차트팩
@@ -58,7 +58,7 @@ namespace Mercury.Backtests.BacktestStrategies
 
 			// 지정가 진입시 진입하는 캔들의 저가가 시가보다 0.03%이상 낮다면 진입 성공
 			// 진입 실패하면 다음 캔들로 넘어가서 진입조건 확인 후 지정가 진입 시도
-			if (trend > 0 && c1.Rsi1 < r.NextDouble(RsiL - RRange, RsiL + RRange))
+			if (trend > 0 && c1.Rsi1 < r.NextDecimal(RsiL - RRange, RsiL + RRange))
 			{
 				var entryPrice = GetAdjustedPrice(PositionSide.Long, c0, true);
 				if (IsOrderFilled(PositionSide.Long, c0, entryPrice, true))
@@ -74,7 +74,7 @@ namespace Mercury.Backtests.BacktestStrategies
 			var c1 = charts[i - 1];
 			var trend = Charts2["BTCUSDT"].GetLatestChartBefore(c1.DateTime.AddDays(-1)).TrendRiderTrend;
 
-			if (trend < 0 && c1.Rsi1 > r.NextDouble(RsiH - RRange, RsiH + RRange))
+			if (trend < 0 && c1.Rsi1 > r.NextDecimal(RsiH - RRange, RsiH + RRange))
 			{
 				var entryPrice = GetAdjustedPrice(PositionSide.Short, c0, true);
 				if (IsOrderFilled(PositionSide.Short, c0, entryPrice, true))
@@ -91,7 +91,7 @@ namespace Mercury.Backtests.BacktestStrategies
 			var c2 = charts[i - 2];
 			// 지정가 청산시 청산하는 캔들의 고가가 시가보다 0.03%이상 높다면 청산 성공
 			// 청산 실패하면 다음 캔들로 넘어가서 청산조건 확인 후 지정가 청산 시도
-			if (c1.Rsi1 > r.NextDouble(RsiL2 - RRange, RsiL2 + RRange))
+			if (c1.Rsi1 > r.NextDecimal(RsiL2 - RRange, RsiL2 + RRange))
 			{
 				var exitPrice = GetAdjustedPrice(PositionSide.Long, c0, false);
 				if (IsOrderFilled(PositionSide.Long, c0, exitPrice, false))
@@ -106,7 +106,7 @@ namespace Mercury.Backtests.BacktestStrategies
 			var c0 = charts[i];
 			var c1 = charts[i - 1];
 			var c2 = charts[i - 2];
-			if (c1.Rsi1 < r.NextDouble(RsiH2 - RRange, RsiH2 + RRange))
+			if (c1.Rsi1 < r.NextDecimal(RsiH2 - RRange, RsiH2 + RRange))
 			{
 				var exitPrice = GetAdjustedPrice(PositionSide.Short, c0, false);
 				if (IsOrderFilled(PositionSide.Short, c0, exitPrice, false))
