+++ Mercury/Backtests/EasyBacktester.cs
@@ -1294,7 +1294,7 @@ namespace Mercury.Backtests
 			return false;
 		}
 
-		bool IsPowerGoldenCross(List<ChartInfo> charts, int lookback, int index, double? currentMacd = null)
+		bool IsPowerGoldenCross(List<ChartInfo> charts, int lookback, int index, decimal? currentMacd = null)
 		{
 			// Starts at charts[index - 1]
 			for (int i = 0; i < lookback; i++)
@@ -1320,7 +1320,7 @@ namespace Mercury.Backtests
 			return false;
 		}
 
-		bool IsPowerGoldenCross2(List<ChartInfo> charts, int lookback, int index, double? currentMacd = null)
+		bool IsPowerGoldenCross2(List<ChartInfo> charts, int lookback, int index, decimal? currentMacd = null)
 		{
 			// Starts at charts[index - 1]
 			for (int i = 0; i < lookback; i++)
@@ -1362,7 +1362,7 @@ namespace Mercury.Backtests
 			return false;
 		}
 
-		bool IsPowerDeadCross(List<ChartInfo> charts, int lookback, int index, double? currentMacd = null)
+		bool IsPowerDeadCross(List<ChartInfo> charts, int lookback, int index, decimal? currentMacd = null)
 		{
 			// Starts at charts[index - 1]
 			for (int i = 0; i < lookback; i++)
@@ -1388,7 +1388,7 @@ namespace Mercury.Backtests
 			return false;
 		}
 
-		bool IsPowerDeadCross2(List<ChartInfo> charts, int lookback, int index, double? currentMacd = null)
+		bool IsPowerDeadCross2(List<ChartInfo> charts, int lookback, int index, decimal? currentMacd = null)
 		{
 			// Starts at charts[index - 1]
 			for (int i = 0; i < lookback; i++)
